backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
dspy-ai==2.4.9
openai==1.12.0
python-dotenv==1.0.0
pydantic==2.5.3
python-multipart==0.0.6
cors==1.0.1